+++ xc3_lib/src/mxmd.rs
@@ -695,8 +695,10 @@ pub struct Texture {
     pub texture_index: u16,
     /// Index into the samplers in [samplers](struct.Materials.html#structfield.samplers).
     pub sampler_index: u16,
-    pub unk2: u16, // TODO: sometimes 1?
-    pub unk3: u16,
+    /// Index into the samplers in [samplers](struct.Materials.html#structfield.samplers).
+    // TODO: This sampler is the same as above but with a float value of 0.0?
+    pub sampler_index2: u16,
+    pub unk3: u16, // 0
 }
 
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]